move DataReceiver to near-vm-errors
See #9176 for the rationale, near-vm-errors will be merged into
near-vm-runner as soon as circular dependencies are done for.
